# perf: improve deserialization of value encoding for struct/list