I hereby agree to the terms of the [Singularity Data, Inc. Contributor License Agreement](https://gist.github.com/skyzh/0663682a70b0edde7ae991492f2314cb#file-s9y_cla).

## What's changed and what's your intention?

- improve deserialization of value encoding for struct/list by preallocating vector. 

The benchmark result is as follows, the deserialization time was reduced by about 40-50%
```shell
bench Struct of Bool (len = 100) (key encoding deserialization)
                        time:   [2.0542 µs 2.0583 µs 2.0625 µs]
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) high mild
  1 (1.00%) high severe

bench Struct of Bool (len = 100) (value encoding deserialization)
                        time:   [1.1362 µs 1.1388 µs 1.1414 µs]
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) high mild

bench List of Bool (len = 100) (key encoding deserialization)
                        time:   [1.1913 µs 1.1944 µs 1.1975 µs]
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild

bench List of Bool (len = 100) (value encoding deserialization)
                        time:   [1.1403 µs 1.1436 µs 1.1476 µs]
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe
```

## Checklist

~- [ ] I have written necessary rustdoc comments~
~- [ ] I have added necessary unit tests and integration tests~
- [x] All checks passed in `./risedev check` (or alias, `./risedev c`)


## Refer to a related PR or issue link (optional)
#5713