@waruto210 2022-10-18
> So almost all `try_collect` in our codebase suffer from the same issue. We lack an "optimistic `try_collect`". 😄

That's too bad, I have seen so many `try_collect` in our codebase.