@waruto210 2022-10-18
> 
> So when we're calling `Result::from_iter`, it won't hit the specialization of initializing with capacity in `Vec::from_iter` (in `spec_from_iter_nested.rs`). Therefore, the performance is worse than manually initializing and pushing in a `for` loop.
> 

There is already an issue here https://github.com/rust-lang/rust/issues/48994